[update]PublicOrders

+++ app/controllers/public/orders_controller.rb
@@ -1,5 +1,5 @@
 class Public::OrdersController < ApplicationController
-  
+
   before_action :authenticate_customer!
 
   def new
@@ -16,12 +16,12 @@ class Public::OrdersController < ApplicationController
       @order.name = current_customer.last_name + current_customer.first_name
 
     elsif params[:order][:address_option] == "1"
-      ship = Address.find(params[:order][:customer_id])
+      ship = Address.find(params[:order][:address_id])
       @order.postal_code = ship.postal_code
       @order.address = ship.address
       @order.name = ship.name
 
-    elsif params[:order][:address_option] = "2"
+    elsif params[:order][:address_option] == "2"
       @order.postal_code = params[:order][:postal_code]
       @order.address = params[:order][:address]
       @order.name = params[:order][:name]
@@ -39,34 +39,36 @@ class Public::OrdersController < ApplicationController
   def create
     @order = Order.new(order_params)
     @order.customer_id = current_customer.id
-    @order.save
 
-    current_customer.cart_items.each do |cart_item|
-      @ordered_item = OrderDetail.new #初期化
-      @ordered_item.order_id =  @order.id
-      @ordered_item.item_id = cart_item.item_id
-      @ordered_item.amount = cart_item.amount
-      @ordered_item.price_taxin = (cart_item.item.price * 1.1).floor
-      @ordered_item.save
+    if @order.save
+      current_customer.cart_items.each do |cart_item|
+        @order_detail = OrderDetail.new(order_id: @order.id)
+        @order_detail.item_id = cart_item.item_id
+        @order_detail.amount = cart_item.amount
+        @order_detail.price = cart_item.item.price
+        @order_detail.save
+      end
+
+      current_customer.cart_items.destroy_all
+      redirect_to complete_public_orders_path
+    else
+      render 'new'
     end
-
-    current_customer.cart_items.destroy_all
-    redirect_to complete_public_orders_path
   end
 
   def index
     @orders = current_customer.orders.all
-    @cart_items = current_customer.cart_items.all
   end
 
   def show
     @order = current_customer.orders.find(params[:id])
-    @cart_items = @order.order_details
+    @order_details = @order.order_details.all
   end
 
   private
 
-    def order_params
-        params.require(:order).permit(:shipping_fee, :payment_method, :name, :address, :postal_code ,:customer_id,:total_payment,:status)
-    end
+  def order_params
+    params.require(:order).permit(:shipping_fee, :payment_method, :name, :address, :postal_code, :customer_id, :total_payment, :status)
+  end
+
 end
